Fixed views.py, accidentally added code into the wrong one.
